Put the creation of class node into its own transformer. BasicSourceSinkManager will try to create a class node if one does not exists for a particular class name by using the system class loader to create a class reader for the class and then visiting the class to record supertype information necessary for the class node.
