+++ Phosphor/src/edu/columbia/cs/psl/phosphor/BasicSourceSinkManager.java
@@ -1,11 +1,15 @@
 package edu.columbia.cs.psl.phosphor;
 
-import edu.columbia.cs.psl.phosphor.instrumenter.TaintTrackingClassVisitor;
 import edu.columbia.cs.psl.phosphor.struct.LinkedList;
 import edu.columbia.cs.psl.phosphor.struct.SimpleHashSet;
+import org.objectweb.asm.ClassReader;
+import org.objectweb.asm.ClassVisitor;
+import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.tree.ClassNode;
 
 import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.Scanner;
 
@@ -66,6 +70,10 @@ public class BasicSourceSinkManager extends SourceSinkManager {
 				set.addAll(originalMethods.get(className));
 			}
 			ClassNode cn = Instrumenter.classes.get(className);
+			if(cn == null) {
+				// Class was loaded before ClassSupertypeReadingTransformer was added
+				cn = tryToAddClassNode(className);
+			}
 			if(cn != null) {
 				if (cn.interfaces != null) {
 					// Add all auto taint methods from interfaces implemented by this class
@@ -80,6 +88,33 @@ public class BasicSourceSinkManager extends SourceSinkManager {
 			}
 			inheritedMethods.put(className, set);
 			return set;
+		}
+	}
+
+	/* Attempts to create a ClassNode populated with supertype information for this class. */
+	private static ClassNode tryToAddClassNode(String className) {
+		try {
+			String resource = className + ".class";
+			InputStream is = ClassLoader.getSystemClassLoader().getResourceAsStream(resource);
+			if (is == null) {
+				return null;
+			}
+			ClassReader cr = new ClassReader(is);
+			cr.accept(new ClassVisitor(Opcodes.ASM5) {
+				@Override
+				public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
+					super.visit(version, access, name, signature, superName, interfaces);
+					ClassNode cn = new ClassNode();
+					cn.name = name;
+					cn.superName = superName;
+					cn.interfaces = new ArrayList<>(Arrays.asList(interfaces));
+					Instrumenter.classes.put(name, cn);
+				}
+			}, ClassReader.SKIP_CODE);
+			is.close();
+			return Instrumenter.classes.get(className);
+		} catch (Exception e) {
+			return null;
 		}
 	}
 
@@ -160,6 +195,10 @@ public class BasicSourceSinkManager extends SourceSinkManager {
 					return curClassName;
 				}
 				ClassNode cn = Instrumenter.classes.get(curClassName);
+				if(cn == null) {
+					// Class was loaded before ClassSupertypeReadingTransformer was added
+					cn = tryToAddClassNode(className);
+				}
 				if(cn != null) {
 					if (cn.interfaces != null) {
 						// Enqueue interfaces implemented by the current class
